Added spot_test :tada:
This map uses a handful of entities to create a spotlight that follows
the players' crosshair, and turns on when he uses the left click.
However, for an unknown reason, the spotlight falls off the pillar it is
standing on when you start using it. To solve it until I (or someone
else) find a fix, just make a few clicks and turns and it should come
back.